Rename Map to MapLibre to avoid conflicts with builtin Map
Resolves #27
