+++ src/lib/index.ts
@@ -12,7 +12,7 @@ export { default as GeolocateControl } from './GeolocateControl.svelte';
 export { default as HeatmapLayer } from './HeatmapLayer.svelte';
 export { default as Layer } from './Layer.svelte';
 export { default as LineLayer } from './LineLayer.svelte';
-export { default as Map } from './Map.svelte';
+export { default as MapLibre } from './MapLibre.svelte';
 export { default as Marker } from './Marker.svelte';
 export { default as MarkerLayer } from './MarkerLayer.svelte';
 export { default as NavigationControl } from './NavigationControl.svelte';
