+++ src/routes/examples/basic/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Map from '$lib/Map.svelte';
+  import MapLibre from '$lib/MapLibre.svelte';
   import CodeSample from '$site/CodeSample.svelte';
   import code from './+page.svelte?raw';
   import type { PageData } from './$types';
@@ -7,10 +7,10 @@
   export let data: PageData;
 </script>
 
-<Map
+<MapLibre
   style="https://basemaps.cartocdn.com/gl/positron-gl-style/style.json"
   class="relative w-full aspect-[9/16] max-h-[70vh] sm:max-h-full sm:aspect-video"
   standardControls
 />
 
-<CodeSample {code} startBoundary="<Map" endBoundary="/>" />
+<CodeSample {code} startBoundary="<MapLibre" endBoundary="/>" />
