+++ src/routes/examples/custom_marker/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Map from '$lib/Map.svelte';
+  import MapLibre from '$lib/MapLibre.svelte';
   import Marker from '$lib/Marker.svelte';
   import { mapClasses } from '../styles';
   import code from './+page.svelte?raw';
@@ -50,7 +50,7 @@
   {/if}
 </p>
 
-<Map
+<MapLibre
   style="https://basemaps.cartocdn.com/gl/positron-gl-style/style.json"
   class={mapClasses}
   standardControls
@@ -72,6 +72,6 @@
       </Popup>
     </Marker>
   {/each}
-</Map>
+</MapLibre>
 
 <CodeSample {code} />
