+++ src/routes/examples/deckgl-arcs/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import MapLibre from '$lib/Map.svelte';
+  import MapLibre from '$lib/MapLibre.svelte';
   import CodeSample from '$site/CodeSample.svelte';
   import code from './+page.svelte?raw';
   import { geoCentroid } from 'd3-geo';
