+++ src/routes/examples/geojson_extrusion/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Map from '$lib/Map.svelte';
+  import MapLibre from '$lib/MapLibre.svelte';
   import GeoJSON from '$lib/GeoJSON.svelte';
   import { mapClasses } from '../styles.js';
   import code from './+page.svelte?raw';
@@ -13,7 +13,7 @@
   Map of US Metropolitan Areas. Height indicates total population and color indicates population
   density.
 </p>
-<Map
+<MapLibre
   style="https://basemaps.cartocdn.com/gl/positron-gl-style/style.json"
   class={mapClasses}
   standardControls
@@ -48,6 +48,6 @@
       </Popup>
     </FillExtrusionLayer>
   </GeoJSON>
-</Map>
+</MapLibre>
 
 <CodeSample {code} />
