+++ src/routes/examples/heatmap/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Map from '$lib/Map.svelte';
+  import MapLibre from '$lib/MapLibre.svelte';
   import GeoJSON from '$lib/GeoJSON.svelte';
   import { mapClasses } from '../styles.js';
   import code from './+page.svelte?raw';
@@ -15,7 +15,7 @@
   >
 </p>
 
-<Map
+<MapLibre
   style="https://basemaps.cartocdn.com/gl/positron-gl-style/style.json"
   class={mapClasses}
   standardControls
@@ -101,7 +101,7 @@
       }}
     />
   </GeoJSON>
-</Map>
+</MapLibre>
 
 <CodeSample {code} />
