+++ src/routes/examples/zoom_transition/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Map from '$lib/Map.svelte';
+  import MapLibre from '$lib/MapLibre.svelte';
   import GeoJSON from '$lib/GeoJSON.svelte';
   import { mapClasses } from '../styles.js';
   import code from './+page.svelte?raw';
@@ -52,7 +52,7 @@
   <p>Current Zoom: {currentZoom.toFixed(1)}</p>
 </div>
 
-<Map
+<MapLibre
   style="https://basemaps.cartocdn.com/gl/positron-gl-style/style.json"
   class={mapClasses}
   standardControls
@@ -133,6 +133,6 @@
       />
     </GeoJSON>
   </ZoomRange>
-</Map>
+</MapLibre>
 
 <CodeSample {code} />
